+++ .env
@@ -8,8 +8,8 @@ CRAWL_EXECUTABLEPATH = /Applications/Google Chrome.app/Contents/MacOS/Google Chr
 # DB_PORT = 3306
 # HOST = localhost
 DB_USERNAME = root
-DB_PASSWORD = 159rjS497m!
-DB_NAME = test1
+DB_PASSWORD = shine2358!
+DB_NAME = db_hasuk
 DB_PORT = 3306
 DB_HOST = localhost
