+++ src/apis/house/dto/createHouse/createHouse.input.ts
@@ -2,6 +2,7 @@ import { Field, InputType, Int } from '@nestjs/graphql';
 import { House_costInput } from 'src/apis/house/dto/createHouse/createHouse.house_cost.input';
 import { House_locationInput } from 'src/apis/house/dto/createHouse/createHouse.house_location.input';
 import { HouseInput } from './createHouse.house.input';
+import { GraphQLUpload, FileUpload } from 'graphql-upload';
 
 @InputType()
 export class CreateHouseInput {
@@ -15,14 +16,12 @@ export class CreateHouseInput {
   house_cost: House_costInput;
 
   @Field(() => Int)
-  university_id: number;
-
-  @Field(() => Int)
   region_id: number;
 
   @Field(() => Int)
   house_category_id: number;
 
-  @Field(() => [String])
-  imgRawDatas: string[]; // 수정 필요
+  @Field(() => [GraphQLUpload], { nullable: true })
+  imgRawDatas: FileUpload[]; // 수정 완료
+
 }
